Adding test folder and moving jest in it.
