+++ tests/jest/snapshots/current_output/welcome-test-tsx-welcome-component-should-match-current-snapshot-1-snap.png
